refactor: restructure tax calculator service and implement health check endpoint
