+++ app/backend/tax-calculator/src/handlers/mod.rs
@@ -1,0 +1,6 @@
+pub mod tax;
+pub mod health;
+
+// Re-export handler functions for convenience
+pub use tax::calculate_tax;
+pub use health::health;
