+++ app/backend/tax-calculator/src/models/error.rs
@@ -1,0 +1,6 @@
+use serde::Serialize;
+
+#[derive(Debug, Serialize)]
+pub struct ErrorResponse {
+    pub error: String,
+}
