+++ app/backend/tax-calculator/src/models/mod.rs
@@ -1,0 +1,7 @@
+pub mod error;
+pub mod invoice;
+pub mod tax;
+
+pub use error::ErrorResponse;
+pub use invoice::{InvoiceData, InvoiceType};
+pub use tax::TaxCalculationResponse;
