+++ app/backend/tax-calculator/src/services/mod.rs
@@ -1,0 +1,1 @@
+pub mod zk_proof;
