+++ app/backend/tax-calculator/src/services/zk_proof.rs
@@ -1,0 +1,12 @@
+/// Placeholder for ZK proof generation
+/// TODO: Implement actual ZK proof generation using Circom circuit
+pub async fn generate_zk_proof(_income: f64, _expenses: f64, _tax_owed: f64) -> bool {
+    println!("Generating ZK proof...");
+    
+    // Simulate proof generation time
+    tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
+    
+    // For now, always return true (proof generated successfully)
+    // Later, this will call our Circom circuit and return actual proof
+    true
+}
